requirements.txt
langchain
python-dotenv
langchain_groq
langchain_community
python-telegram-bot